% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_expression.R
\name{ClusterExons}
\alias{ClusterExons}
\title{Detect constitutive and alternative exons}
\usage{
ClusterExons(data = NULL, exonID = NULL, exceptions = ceiling(ncol(data) *
  0.1), signDigits = 3, distMethod = "correlation", link = "complete",
  nbproc = 1)
}
\arguments{
\item{data}{A data.frame or numeric matrix with one row per exon and one column per sample, containing the expression values of each exon.}

\item{exonID}{A data.frame with the following columns. \describe{
\item{geneID}{Gene ID of each exon.}
\item{exonID}{ID of each exon.}
}}

\item{exceptions}{Integer. Sets the maximum number of samples in which the constitutive exon group does not need to be the most expressed one. Defaults to 10\% of the dataset.}

\item{signDigits}{An integer. Number of digits rounded from expression scores for ranking calculations.}

\item{distMethod}{The distance metric used by the clustering algorithm (default: correlation), see function \code{\link[amap:hcluster]{hcluster}} from package amap for more information.}

\item{link}{Agglomeration method used by the clustering algorithm (default: complete), see function \code{\link[amap:hcluster]{hcluster}} from package amap for more information.}

\item{nbproc}{Integer. Number of subprocess for parallelization (default: 1), see function \code{\link[amap:hcluster]{hcluster}} from package amap for more information}
}
\value{
A data.frame with the following columns. \describe{
\item{geneID}{Gene ID of each exon.}
\item{clusters}{Numeric ID for each exon group detected.}
\item{exonID}{Exon ID.}
\item{splicing_category}{Gene class. Can be either spliced (isoforms detected), unspliced (no isoforms detected) or  single_expressed_exon (only one exon present in the gene).}
\item{constitutive}{Identifies whether the exon group is constitutive (main gene) or facultative (isoform).}
\item{transcriptID}{Identifier for each exon group detected. Ends in '_con' for constitutive and '_fac' for facultative exons.}
}
}
\description{
Cluster genes according to reciprocal correlations, then iteratively cut tree (bottom up)
until one cluster is the most expressed or tied for expression across all samples.
Uses \code{\link[amap:hcluster]{hcluster}} from amap package for clustering
}
