\name{Example}
\docType{data}
\alias{Example}
\title{Time series of commercial fish landings and fishing effort along Andhra Pradesh coastline (1997-2018)}
\description{A list named "Example" has been taken for illustration. It contains three data frames named sp_catch (Quantity of the fish species, in tonnes), tot_catch (Quantity of total catch, in tonnes) and effort (Fishing duration, in hours) with the same dimension. Data was collected from the commercial fish landing centres along the coast of A.P. through a scientifically and systematically planned sampling design, employing the stratified multi-stage random sampling technique. The time series (1997-2018) of catch of Indian Mackerel, the total catch and fishing effort expended by 8 different fishing gears viz., mechanized trawl net including multiday trawl net (MTN), mechanized gill net (MGN), non-mechanized gears (NM), outboard gill net (OBGN), outboard ring seine (OBRS), outboard trawl net (OBTN), some minor mechanized gears (MOTHERS) and some minor outboard gears (OBOTHS) have been used.

}
\usage{Example}

\format{Each data frame is having 22 rows and 9 variables}

\source{Time series of catch (in tonnes) and effort (in hours of operation) from 1997 to 2018 was taken from National Marine Fishery Resources Data Centre (NMFDC) of CMFRI, Kochi}

\references{
http://www.cmfri.org.in/fish-catch-estimates
}

\keyword{datasets}

