% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_testdata.R
\name{genTestdata}
\alias{genTestdata}
\title{Create test dataset for the ClusterExons function}
\usage{
genTestdata(nGenes, nExons, nSamples)
}
\arguments{
\item{nGenes}{An integer, specifies the number of genes in the dataset.}

\item{nExons}{An integer, specifies the number of exons in each gene.}

\item{nSamples}{An integer, specifies the number of samples in the dataset.}
}
\value{
A data.frame with nGenes*nExons rows and 2+nSamples columns. \newline
The first 2 columns contain gene and exon ID
Further columns contain log2 transformed expression scores
}
\description{
This function creates a sample test dataset for the ClusterExons function.
At the moment it the same number of exons per each gene.
Expression is sampled from a random binomial distribution.
}
\examples{
# Create a dataset with 10 genes composed by 3 exons, measured across 5 samples
genTestdata(10,3,5)

 @export
}
