% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_expression.R
\name{AverageExons}
\alias{AverageExons}
\title{Average expression values based on unique eigenexon IDs}
\usage{
AverageExons(data = NULL, spliceID = NULL, splicingRatios = F,
  NAcorrection = F)
}
\arguments{
\item{data}{A data.frame or numeric matrix with one row per exon and one column per sample, containing the expression values of each exon.}

\item{spliceID}{A data.frame. Output from the \code{\link{ClusterExons}} function.}

\item{splicingRatios}{Logical. If FALSE, expression from all entries is reported on the same scale. If TRUE, expression from splicing entries is normalized by their gene's constitutive expression score, generating splicing ratios.}

\item{NAcorrection:}{Logical. Applicable only if splicingRatios is TRUE. If TRUE, splicing ratios higher than 1 are set to 1 and NA/NaN/infinity values to 0. This accounts for experimental error in measurements.}
}
\description{
Average expression values based on unique eigenexon IDs
}
